Redux Refactoring
